+++ account/messages.py
@@ -1,0 +1,36 @@
+from django.utils.translation import gettext_lazy as _
+
+PERMISSION_NOT_ALLOWED = _('Você não tem permissão para executar essa ação')
+INVALID_PASSWORD = _('Senha inválida')
+FOREIGN_KEY_EXCEPTION = _('Registro com dependências e não pode ser excluído')
+SETTINGS_EMAIL_NOT_FOUND = _('Configurações de E-mail não encontrado')
+INVALID_SECRET_KEY = _('Senha inválida!')
+ERROR_TO_SEND_EMAIL = _('Erro ao enviar e-mail!')
+INVALID_EMAIL = _('Entre com um email valido')
+USER_NOT_FOUND = _('Usuário não encontrado!')
+USER_INVALID = _('Email e telefone não correspondem a nenhum usuário!')
+VALID_PASSWORD = _('Senha alterada com sucesso!')
+NO_RECORD_FOUND = _('Nenhum registro encontrado!')
+REPORT_SERVER_UNAVAILABLE = _('Servidor de relatório indisponível')
+PROJECT_EXISTS = _('O nome do projeto já existe!')
+ACTION_FAILED = _('Não foi possível executar esta ação, causa: %s')
+NO_RECORDS_FOUND = _('Nenhum registro encontrado')
+MINIO_FAILURE = _('Falha de comunicação com Minio')
+GROUP_DOES_NOT_EXISTS = _('Este grupo não existe!')
+NO_ACTIVE_VOTING = _('Sem votações ativas no momento')
+USER_HAS_ALREADY_VOTED = _('Esse usuário já votou nessa eleição!')
+EMAIL_ALREADY_EXISTS = _('Esse e-mail já está sendo usado por outro usuário!')
+TIME_OVER_VOTING = _('Desculpe, a votação já ultrapassou o tempo limite estabelecido.')
+INVALID_CREDENTIALS = _('Credenciais inválidas!')
+CLOSE_VOTE = _('Votação finalizada com sucesso!')
+ACTIVE_VOTE = _('Votação ativada novamente!')
+INVALID_DELETE_VOTE = _('Não é possivel deletar votação com chapa ativa!')
+EXISTS_VOTE_ACTIVE = _('Não é possível ativar essa votação, pois existe outra votação ativa!')
+UNABLE_REMOVE_USER = _('Não foi possível remover esse usuário!')
+UNABLE_REMOVE_PLATE = _('Não foi possível remover essa chapa!')
+CELLPHONE_ALREADY_EXISTS = _('Esse numéro de celular já está sendo usado por outro usuário!')
+PLATE_ALREADY_EXISTS = _('Já existe uma chapa cadastrada com o mesmo nome!')
+VOTING_ALREADY_EXISTS = _('Já existe uma votação cadastrada com o mesmo nome!')
+CELLPHONE_DOES_NOT_USER = _('O celular cadastrado não está vinculado a nenhum usuário!')
+EVENT_VOTING_ERROR = _('Não e possivel atualizar a votação')
+PLATE_USER_ACTIVE = _('Os candidatos dessa chapa já está em outra chapa ativa!!')
