+++ core/serializers.py
@@ -1,46 +1,22 @@
-from django.contrib.auth.hashers import make_password
 from rest_framework import serializers
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-from core import models, utils
-from rest_flex_fields import FlexFieldsModelSerializer
+
+from account import serializers as account_serializers
+from core import models
 
 
-class SerializerBase(FlexFieldsModelSerializer, serializers.HyperlinkedModelSerializer):
-
-    def get_field_names(self, declared_fields, info):
-        fields = super().get_field_names(declared_fields, info)
-        fields.insert(0, 'id')
-        return fields
-
-
-class UserSerializer(SerializerBase):
-    def validate_password(self, value: str) -> str:
-        """
-        Hash value passed by user.
-
-        :param value: password of a user
-        :return: a hashed version of the password
-        """
-        return make_password(value)
-
-    class Meta:
-        model = models.User
-        fields = '__all__'
-
-
-class VoterSerializer(SerializerBase):
+class VoterSerializer(account_serializers.SerializerBase):
     class Meta:
         model = models.Voter
         fields = '__all__'
 
 
-class CandidateSerializer(SerializerBase):
+class CandidateSerializer(account_serializers.SerializerBase):
     class Meta:
         model = models.Candidate
         fields = '__all__'
 
 
-class PlateSerializer(SerializerBase):
+class PlateSerializer(account_serializers.SerializerBase):
     was_voted = serializers.BooleanField(read_only=True)
 
     class Meta:
@@ -55,7 +31,7 @@ class PlateSerializer(SerializerBase):
     }
 
 
-class EventVotingSerializer(SerializerBase):
+class EventVotingSerializer(account_serializers.SerializerBase):
     was_voted = serializers.BooleanField(read_only=True)
 
     class Meta:
@@ -63,7 +39,7 @@ class EventVotingSerializer(SerializerBase):
         fields = '__all__'
 
 
-class PlateUserSerializer(SerializerBase):
+class PlateUserSerializer(account_serializers.SerializerBase):
     class Meta:
         model = models.PlateUser
         fields = '__all__'
@@ -76,7 +52,7 @@ class PlateUserSerializer(SerializerBase):
     }
 
 
-class PlateUserPresidentsOrViceSerializer(SerializerBase):
+class PlateUserPresidentsOrViceSerializer(account_serializers.SerializerBase):
     candidate = CandidateSerializer()
     plate = PlateSerializer()
 
@@ -85,7 +61,7 @@ class PlateUserPresidentsOrViceSerializer(SerializerBase):
         fields = '__all__'
 
 
-class VotingPlateSerializer(SerializerBase):
+class VotingPlateSerializer(account_serializers.SerializerBase):
     class Meta:
         model = models.VotingPlate
         fields = '__all__'
@@ -102,13 +78,13 @@ class VotingPlateSerializer(SerializerBase):
     }
 
 
-class VotingUserSerializer(SerializerBase):
+class VotingUserSerializer(account_serializers.SerializerBase):
     class Meta:
         model = models.VotingUser
         fields = '__all__'
 
 
-class ResumeVoteSerializer(SerializerBase):
+class ResumeVoteSerializer(account_serializers.SerializerBase):
     class Meta:
         model = models.ResumeVote
         fields = '__all__'
@@ -123,25 +99,3 @@ class ResumeVoteSerializer(SerializerBase):
             {'fields': ['id', 'description']}
         )
     }
-
-
-class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
-    @classmethod
-    def get_token(cls, user):
-        token = super().get_token(user)
-        token['user_id'] = user.id
-        token['username'] = user.username
-        token['email'] = user.email
-        token['user'] = utils.get_user(user)
-
-        return token
-
-    def validate(self, attrs):
-        data = super().validate(attrs)
-        refresh = self.get_token(self.user)
-        data['refresh'] = str(refresh)
-        data['access'] = str(refresh.access_token)
-        return {
-            'token': data,
-            'user': utils.get_user_login(self.user)
-        }
